requirements.txt
ai2thor==3.1.0
boto3==1.15.2
fuzzywuzzy==0.18.0
networkx==2.5
pydub==0.24.1
python-Levenshtein
tqdm
pillow
pydantic==1.8.2
numpy
pandas
opencv-python
vocab
revtok
Pillow
sacred
etaprogress
scikit-video
lmdb
gtimer
filelock
termcolor
torch==1.7.1
torchvision==0.8.2
tensorboardX==1.8
Flask
flask_restful
future
pandoc
six
typing
ConfigArgParse